float wartezeit LEd fix 1

+++ gui.py
@@ -35,11 +35,11 @@ ArrayErstellen()
 
 def los(event):
     print('c', column)
-    zeitzwspalten = int(entry_time.get()) / width
+    zeitzwspalten = entry_time.get() / width
     print(zeitzwspalten)
 
     if platform == "linux" or platform == "linux2":
-        hardware.LEDAnzeigen(column, width, height, pixels, int(zeitzwspalten))
+        hardware.LEDAnzeigen(column, width, height, pixels, float(zeitzwspalten))
 
     elif platform == "win32":
